TheGreyPlanet/requirements.txt
asgiref==3.2.3
chardet==3.0.4
Django==3.0.3
django-etc==1.2.0
django-summernote==0.8.11.6
gunicorn==20.0.4
ldif3==3.1.1
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-card-me==0.9.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
surlex==0.2.0
text-unidecode==1.3
unicodecsv==0.14.1
